+++ libs/base_core/src/media/image/mod.rs
@@ -6,13 +6,13 @@
 // #![cfg_attr(base_safe_image, forbid(unsafe_code))] // no feature
 
 // mod pnm;
-// pub mod sixel; // SixelChar, SixelColor, SixelEncoder, SixelPalette
+pub mod sixel; // SixelChar, SixelColor, SixelEncoder, SixelPalette
 
 crate::structural_mods! { // _pub_mods
     _pub_mods {
-        // pub use super::{
-        //     // pnm::*,
-        //     // sixel::*,
-        // };
+        pub use super::{
+            // pnm::*,
+            sixel::_all::*,
+        };
     }
 }