+++ src/media/image/sixel/mod.rs
@@ -17,6 +17,11 @@ const SIXEL_PALETTE_MAX: usize = 256;
 
 crate::structural_mods! { // _mods
     _mods {
+        // re-exports
+        pub use devela_base_core::media::image::{
+            SixelChar, SixelColor, SixelEncoder, SixelPalette,
+        };
+
         pub use super::{builder::*, dither::*, error::*, output::*, quant::*};
     }
 }